+++ src/components/molecules/CookieDeclaration/CookieDeclaration.tsx
@@ -1,7 +1,20 @@
+import { useEffect } from "react";
+import { useTranslation } from "react-i18next";
 import { ArrorDown } from "src/components/atoms/icons/ArrowDown/ArrowDown";
 import { World } from "src/components/atoms/icons/World/World";
 
+const lngs = {
+  en: { nativeName: "English" },
+  pt: { nativeName: "Portugues" },
+};
+
 const CookieDeclaration = () => {
+  const { i18n } = useTranslation();
+
+  useEffect(() => {
+    i18n.changeLanguage("en");
+  }, []);
+
   return (
     <section className="flex flex-col items-center">
       <a
@@ -22,10 +35,13 @@ const CookieDeclaration = () => {
         <select
           name="languages"
           className="appearance-none mb-7 pr-3 bg-yellow font-bold text-lg border-b border-black hover:cursor-pointer"
+          onChange={(e) => i18n.changeLanguage(e.target.value)}
         >
-          <option value="english">English</option>
-          <option value="portuguese">Portuguese</option>
-          <option value="spanish">Spanish</option>
+          {Object.keys(lngs).map((lng) => (
+            <option value={lng}>
+              {lngs[lng as keyof typeof lngs].nativeName}
+            </option>
+          ))}
         </select>
 
         <ArrorDown className="absolute right-0 top-2.5" />
