+++ src/components/molecules/Medias/Medias.tsx
@@ -1,10 +1,11 @@
-import { Tumblr } from "@molecules";
+import { useTranslation } from "react-i18next";
 import { Discord } from "src/components/atoms/icons/Discord/Discord";
 import { Facebook } from "src/components/atoms/icons/Facebook/Facebook";
 import { Forums } from "src/components/atoms/icons/Forums/Forums";
 import { Instagram } from "src/components/atoms/icons/Instagram/Instagram";
 import { Twitch } from "src/components/atoms/icons/Twitch/Twitch";
 import { Twitter } from "src/components/atoms/icons/Twitter/Twitter";
+import { Tumblr } from "src/components/atoms/icons/Tumblr/Tumblr";
 import { Youtube } from "src/components/atoms/icons/Youtube/Youtube";
 
 const socialMedias = [
@@ -51,9 +52,11 @@ const socialMedias = [
 ];
 
 const Medias = () => {
+  const { t } = useTranslation();
+
   return (
     <>
-      <span className="font-bold mb-2">Find us on</span>
+      <span className="font-bold mb-2">{t("findus")}</span>
       <ul className="flex flex-wrap justify-around w-full p-3">
         {socialMedias.map(({ name, component, link }) => (
           <li
